Rework again the !barbut command
